requirements.txt
telethon
aiogram
google-genai
python-dotenv
pydantic-settings
structlog
httpx
beautifulsoup4
